requirements.txt
# DriftGuard Python Dependencies
pydantic>=2.0.0
boto3>=1.26.0
kubernetes>=26.1.0
PyYAML>=6.0
requests>=2.28.0
prometheus-client>=0.20.0
confluent-kafka>=2.0.0